Modified price formation algorithm
The results should be more or less the same as before, but much faster. Before
it ran in roughly O(N^2), where N is the number of agents. Now is runs in
roughly O(N log N), as the price formation iterates through a sorted array.

However, in rare edge cases (many agents bid identical prices, but not all
orders can be met), this degrades to O(N*A), where A is the number of assets
per agent. Talk to daniel if you encounter this scenario.

Also made all the agents pay the market clearing price (pstar), so that money
doesn't enter/exit the system.